+++ sdk/examples/word/Utility.cpp
@@ -3,5 +3,5 @@
 float _ceilf(float f)
 {
     float i = (float)(int)f;
-    return (fmodf(f, 1.f)) ?  i + 1.f : i;
+    return (fmod(f, 1.f)) ?  i + 1.f : i;
 }
